+++ src/BinaryString.cpp
@@ -222,7 +222,7 @@ QByteArray BinaryString::value() const {
  */
 void BinaryString::setValue(const QByteArray &data) {
 
-	valueOriginalLength_ = data.size();
+	valueOriginalLength_ = static_cast<int>(data.size());
 	on_keepSize_stateChanged(ui->keepSize->checkState());
 	const auto temp = QString::fromLatin1(data.data(), data.size());
